requirements.txt
line-bot-sdk
gunicorn==19.9.0
numpy
flask